Split IHA steps in apply/undo with vars to exclude
This commit adds the undo procedure for the Instance HA role. Now by
default the role just applies the steps for the IHA enablement, but it
is possible to override the behavior with the vars apply_instance_ha and
undo_instance_ha to include or exclude a sequence of steps.

Change-Id: I24454b29bbf032ac4c4149851175a973e00bb9b8