Gather facts also before validate-ha
Since we use the variable ansible_hostname when cycling on the inventory
hosts file, we need to gather_facts from all the overcloud nodes,
otherwise will get an error related to the undefined variable.
This also force us to split the way in which we copy ha-test-suite on
undercloud and overcloud nodes, since the undercloud node can have an
ansible_hostname different from "undercloud".

Change-Id: I3af2895110ccd89209feaa3da0a1b35d6842758e