Force failure after ops in rescue and fix wait opt
Today we manage tests between "blocks:" with "always:" and "rescue:" to
get the output of the commands, both stderr and stdout. Problem is that
instead of stopping we move on, and this is not useful when you want to
investigate on problems like the --wait 600 versus --wait=600 (that this
patch solves, too).

Change-Id: Icd5f6f8bdff8c563131856286e82f9a7169b5fc7