Fix Rally OSP12/Pike support for IHA plugin
This commit fix the way the Rally plugin looks for the hypervisor IP
that is running the instance inside OSP12/Pike deployments. Without this
the IP would be resolved but with an address (default is the internal
network one) not reachable by the undercloud from which we need to
launch the failover command.

Change-Id: I93db75b72563f191cac73a4a749827ed3935493b