Fix Pike/OSP12 IHA cluster wide setting
This commit adds an essential command to the apply steps for the
Instance HA in Pike/OSP12:

    pcs resource defaults requires=fencing

Without this in place once the compute's unfence trigger is created all
the bundle resources are stopped.
By default [1] "requires" meta attribute for each resource is calculated
depending on the status of the other resources. It should be "fencing"
since we have stonith-enabled to true, but since we are creating also a
compute dummy resource with provides=unfencing then it seems that the
requires defaults become "unfencing" too, so we need to explicit set
the default for all the other resources.

[1] http://clusterlabs.org/doc/en-US/Pacemaker/1.1/html-single/Pacemaker_Explained/#_resource_meta_attributes

Change-Id: I3a93c3c8ecf9bb9e4d2fb1ca2cecaf777e7547aa