Add overcloudrc.v3 support in IHA
This commit adds the support to the stack rc file related to the version
3 of keystone. It will check for an overcloudrc.v3 file in the
working_dir of the deployment (typically /home/stack) and if it is not
able to find it it will use overclourc as usual.
The command for creating the nova-evacuate resource was changed
accordingly, adding --force to support liberty/osp8 and mitaka/osp9
because in these releases we can't provide the new requested fields
project_domain and user_domain.
The way overcloudrc was treated before has been cleaned. No need to copy
the file on the controller node. The name of the file is calculated
dynamically as before, also for v3.

Note: this commit solves also what was proposed here [1] when dealing
with overcloudrc files in which there are both OS_TENANT_NAME and
OS_PROJECT_NAME variables, taking the latter.
The stonith creation python script has been modified to support also
Ocata, for which we still have OS_TENANT_NAME but we don't support
positional arguments anymore.

Change-Id: I1a143b67850c67d6d3207c06abcf60d32f2456ff