Fix the way iptables rules are managed
Today we use iptables-save to store the configuration related to the
firewall (we enable INPUT traffic for pacemaker remote on all the hosts)
but this method is not reliable since Neutron put and remove rules by
itself, and by using iptables-save we also store them.
This patch treat the single rule in both apply and undo playbooks so
that we don't touch anything that neutron might touch.

Change-Id: I6293d7a065f8e531de6218f272a8b08844b3eb42