Fix custom STONITH compute names in IHA playbook
Today when STONITH for the compute nodes is already configured in the
TripleO deployment and the name of these STONITH resources is not
"ipmilan-<hostname>" the task will likely fail.
This commit fixes the way STONITH levels are managed, making the task to
not rely anymore on the fact that a user will configure also STONITH
with the instance HA playbooks, but reading dynamically the name of the
resources.
It also takes care about it while removing the IHA steps, preserving the
idempotent principle.

Closes-Bug: #1780434

Change-Id: I2f8351cf62cddd420c257110d60442ca4243a43d