Fix hard-coded domain in instance HA role
This commit sets a default value of "localdomain" for the domain
parameter related to the overcloud.
This is used inside the role while configuring nova-compute-checkevacuate
resource and the fence-nova STONITH device. In this way, the user will
be able to override the default by passing the overcloud_domain
parameter in the command line.

Fixes-Bug: #1782752
Change-Id: Ib76523a08e43b8d9f604e9fb2544017c9800d8a9