Apply/Delete latency during HA validation
This commit adds the latency management into the HA validation process.
If a user wants to go trough the HA validation with a specific latency
applied on all the nodes he will pass the "latency_ms" variable with a
value greater than zero and a "latency_eth_interface" variable with the
value related to the device chose for the test.

Change-Id: I3d0f9cc3301ba6be893919b58473f3f4135ea789